+++ lib/melange_edn.mli
@@ -6,7 +6,7 @@ type vector
 type list_
 type number
 
-type _ t =
+type _ t = private
   | Nil : unit t
   | Bool : bool -> bool t
   | String : string -> string t
@@ -26,6 +26,23 @@ type _ t =
 and any = Any : _ t -> any
 
 exception Parse_error of string
+
+val any : _ t -> any
+val nil : unit t
+val bool : bool -> bool t
+val string : string -> string t
+val char : Uchar.t -> Uchar.t t
+val symbol : string -> symbol t
+val keyword : string -> keyword t
+val int : int64 -> number t
+val bigint : string -> number t
+val float : float -> number t
+val decimal : string -> number t
+val list : any list -> list_ t
+val vector : any list -> vector t
+val map : (any * any) list -> map t
+val set : any list -> set t
+val tagged : string -> any -> (string * any) t
 
 val of_edn_string : string -> any
 val of_edn_string_all : string -> any list